+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
-## [Version 1.0.1](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.0.1) - Feature release - 2022-11-21
+## [Version 1.0.1](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.0.1) - Feature release - 2023-01-05
 
 - Add batch mode for metric retrieval with the event frames
 - Disable by default server URL override
 - Fix pagination for Event Frames Search connector
 - Fix Transpose & Synchronize recipe with rows containing no timestamp
 - Fix nested value object when batch mode is not activated on Event Frames Search connector
+- Fix presence of NaNs in last line produced by Transpose & Synchronize recipe
+- Fix Event Frame WebId consistency in batch mode
 
 ## [Version 1.0.0](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.0.0) - Initial release - 2022-08-04